+++ tests/unit/test_main_api.py
@@ -1,7 +1,4 @@
-import os
-from pathlib import Path
 
-import pytest
 
 from splurge_unittest_to_pytest import convert_string, convert_file
 from splurge_unittest_to_pytest.main import is_unittest_file, find_unittest_files